requirements.txt
# COMPASS Backend Dependencies
# Python 3.12+

# ---- Google Cloud -------------------------------------------------------
google-genai>=1.0.0
google-adk>=0.4.0
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.16.0
google-cloud-aiplatform>=1.57.0
vertexai>=1.0.0

# ---- FastAPI / ASGI ------------------------------------------------------
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9         # file upload support
websockets>=12.0

# ---- Data models ---------------------------------------------------------
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ---- Auth ----------------------------------------------------------------
python-jose[cryptography]>=3.3.0
firebase-admin>=6.5.0

# ---- Async helpers -------------------------------------------------------
httpx>=0.27.0
aiofiles>=23.2.1

# ---- Testing -------------------------------------------------------------
pytest>=8.2.0
pytest-asyncio>=0.23.7
pytest-mock>=3.14.0

# ---- Linting / formatting (dev) ------------------------------------------
ruff>=0.4.0
